Tile map moves with player

+++ Game/Engine Source Code/Source Files/TileMap.cpp
@@ -24,12 +24,12 @@ void TileMap::LoadMap(std::string mapPath, const char* spriteSheetPath, Vector2 
 		for (int x = 0; x < mapSize.x; x++)
 		{
 			mapFile.get(c);
-			sourceY = atoi(&c) * tileSize.y;
+			sourceY = static_cast<int>(atoi(&c) * tileSize.y);
 
 			mapFile.get(c);
-			sourceX = atoi(&c) * tileSize.x;
+			sourceX = static_cast<int>(atoi(&c) * tileSize.x);
 
-			AddTile(spriteSheetPath, Vector3(x * tileSize.x * scale, y * tileSize.y * scale, 0), tileSize, Vector2(sourceX, sourceY), scale);
+			AddTile(spriteSheetPath, Vector3(x * tileSize.x * scale, y * tileSize.y * scale, 0), tileSize, Vector2(static_cast<float>(sourceX), static_cast<float>(sourceY)), scale);
 			mapFile.ignore();
 		}
 	}
